mgr/influx: Use Queues to process statistics
By using Queues we can use a multithreaded async method to send
all statistics to InfluxDB instead of using a single-thread in a
for loop.

Deployments on ~2.000 OSD clusters showed that it could take up to
20 seconds to gather and process al statistics and send them to
InfluxDB.

By using a combination of Generators and a Queue we can send out
statistics as we gather them and send them towards InfluxDB faster
and more efficient.

Signed-off-by: Wido den Hollander <wido@42on.com>